+++ CHANGELOG.md
@@ -14,6 +14,8 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Changed
 
+- Made TaxonName autocmlete exact = true by default.
+
 
 ### Fixed